requirements/requirements-docs.txt
appdirs
babel
backrefs
beautifulsoup4
brotli
certifi
charset-normalizer
click
colorama
filelock
gdown
ghp-import
griffe
idna
inflate64
jinja2
joblib
markdown
markupsafe
mergedeep
mkdocs
mkdocs-autorefs
mkdocs-get-deps
mkdocs-material
mkdocs-material-extensions
mkdocs-section-index
mkdocstrings
mkdocstrings-python
multivolumefile
mymodule
numpy
packaging
paginate
pandas
pathspec
platformdirs
psutil
py7zr
pybcj
pycryptodomex
pygments
pymdown-extensions
pyppmd
pysocks
python-dateutil
pytz
pyyaml
pyyaml-env-tag
pyzstd
requests[socks]
scikit-learn
scipy
six
soupsieve
texttable
threadpoolctl
tqdm
typing-extensions ; python_version < "3.11"
tzdata
urllib3
watchdog